Refactored and cleaned up code
Cleaned up main, and cleaned up some code. Still needs some work though.

Fixed a bug when logging out and relogging in that would cause a crash.

Also discovered a bug with the enrollment and disenrollment functions.
If you enroll in a course then immediately try to disenroll the course
will not be found. You need to exit the program then reopen it for it to
work.